requirements.txt
networkx
pyyaml
pandas
torch
numpy
simpy
matplotlib
streamlit
